pyproject.toml
[project]
name = "bilancio"
version = "0.1.0"
description = "Personal finance tracker"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "aiosqlite",
    "alembic",
    "pydantic-settings",
    "structlog",
    "argon2-cffi",
    "slowapi",
    "python-multipart",
    "openpyxl>=3.1.5",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "ruff",
    "mypy",
    "sqlalchemy[mypy]",
    "types-pyyaml>=6.0.12.20260408",
    "types-openpyxl>=3.1.5.20260408",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bilancio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]
omit = ["*/migrations/*", "*/alembic/*"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "B008",  # Depends() in function defaults is the FastAPI idiom
    "E501",  # line length enforced by ruff format, not lint
]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
